requirements.txt
# OMNIMIND LOCAL — Python Dependencies

# Core
fastapi>=0.104
uvicorn[standard]>=0.24
websockets>=12.0
redis>=5.0
pyyaml>=6.0
python-dotenv>=1.0
httpx>=0.25
aiohttp>=3.9
aiofiles>=23.0

# LLM
llama-cpp-python>=0.2.50

# STT
faster-whisper>=1.0

# TTS
piper-tts>=1.2

# Wake Word + VAD
openwakeword>=0.6
# silero-vad installed via torch.hub

# Memory / RAG
chromadb>=0.4
sentence-transformers>=2.3
rank-bm25>=0.2
whoosh>=2.7

# ML / Training
torch>=2.1
transformers>=4.36
datasets>=2.16
accelerate>=0.25
unsloth>=2024.1
trl>=0.7

# Audio
pyaudio>=0.2.14
sounddevice>=0.4
numpy>=1.24
scipy>=1.11

# Agents
python-obd>=0.7
caldav>=1.3

# Security
presidio-analyzer>=2.2
presidio-anonymizer>=2.2
cryptography>=41.0

# Monitoring
psutil>=5.9
prometheus-client>=0.19

# Dev
pytest>=7.4
pytest-asyncio>=0.23
ruff>=0.1

# Drone
# djitellopy>=2.5        # Tello SDK (uncomment when needed)
# pymavlink>=2.4         # MAVLink protocol (uncomment when needed)

# Robotics
pyserial>=3.5             # Serial communication with Arduino/ESP32

# Vision (shared by drone + surveillance)
ultralytics>=8.1          # YOLOv8
opencv-python>=4.8        # Camera capture + image processing
